requirements.txt
django
pymongo[snappy,gssapi,srv,tls]
dnspython
django-allauth
python-dotenv
requests